split into desktop and mobile component (#945)
